twin_recognition_app/requirements.txt
streamlit
opencv-python
numpy
Pillow
deepface